requirements.txt
asgiref==3.9.1
certifi==2025.8.3
contourpy==1.3.2
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.5
fonttools==4.58.5
gunicorn==23.0.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.3
oauthlib==3.3.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.1
sentry-sdk==2.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.9.0
xyzservices==2025.1.0
gunicorn>=21.2

